auth done and mapped with notes

+++ src/main/kotlin/com/example/spring_practice/controller/NotesController.kt
@@ -1,23 +1,17 @@
-package com.example.spring_practice
+package com.example.spring_practice.controller
 
-import com.example.spring_practice.NotesController.NoteResponse
 import com.example.spring_practice.model.Note
 import com.example.spring_practice.repositories.NoteRepo
 import org.bson.types.ObjectId
-import org.springframework.web.bind.annotation.DeleteMapping
-import org.springframework.web.bind.annotation.GetMapping
-import org.springframework.web.bind.annotation.PathVariable
-import org.springframework.web.bind.annotation.PostMapping
-import org.springframework.web.bind.annotation.RequestBody
-import org.springframework.web.bind.annotation.RequestMapping
-import org.springframework.web.bind.annotation.RequestParam
-import org.springframework.web.bind.annotation.RestController
+import org.springframework.security.core.context.SecurityContextHolder
+import org.springframework.web.bind.annotation.*
 import java.time.Instant
 
 
 // post http://localhost:8085/notes
 //get http://localhost:8085/notes?ownerId
 //delete http://localhost:8085/notes/123
+
 
 @RestController
 @RequestMapping("/notes")
@@ -48,14 +42,15 @@ class NotesController(private val noteRepo: NoteRepo) {
     @PostMapping
     fun save(
         @RequestBody body : NoteRequest):NoteResponse{
+        val ownerId = SecurityContextHolder.getContext().authentication.principal as String
         val response =  noteRepo.save(
             Note(
-                id = body.id?.let { ObjectId(it)} ?: ObjectId.get(),
+                id = body.id?.let { ObjectId(it) } ?: ObjectId.get(),
                 title = body.title,
                 content = body.content,
                 color = body.color,
                 createdAt = Instant.now(),
-                ownerId = ObjectId()
+                ownerId = ObjectId(ownerId)
 //                ownerId = ObjectId(body.ownerId)
             )
         )
@@ -64,8 +59,8 @@ class NotesController(private val noteRepo: NoteRepo) {
 
     @GetMapping
     fun findByOwnerId(
-        @RequestParam(required = true) ownerId: String
     ) : List<NoteResponse>{
+        val ownerId = SecurityContextHolder.getContext().authentication.principal as String
         return noteRepo.findByOwnerId(ownerId = ObjectId(ownerId)).map {
             it.toResponse()
         }
@@ -74,16 +69,24 @@ class NotesController(private val noteRepo: NoteRepo) {
 
     @DeleteMapping(path = ["/{id}"])
     fun deleteById(@PathVariable id: String?){
-        val response = noteRepo.deleteById(ObjectId(id))
-    }
-}
+        val note= noteRepo.findById(ObjectId(id)).orElseThrow{
+            IllegalArgumentException("wrong id")
+        } // id is being used to get the ownerid from db
 
-private fun Note.toResponse(): NotesController.NoteResponse{
-    return NoteResponse(
-        id = id.toHexString(),
-        title = title,
-        color = color,
-        content = content,
-        createdAt = createdAt
-    )
+        val ownerId = SecurityContextHolder.getContext().authentication.principal as String // ownweid from the system when user was created
+
+        if(note.ownerId.toHexString() == ownerId)
+        noteRepo.deleteById(ObjectId(id))
+    }
+
+    private fun Note.toResponse(): NotesController.NoteResponse{
+        return NoteResponse(
+            id = id.toHexString(),
+            title = title,
+            color = color,
+            content = content,
+            createdAt = createdAt
+        )
+    }
+
 }
